Automated g4 rollback of changelist 128621412.
*** Reason for rollback ***

TAP has detected that 500 or more targets failed to build at [] each of which successfully built at the prior CL that affected them.
TAP double-checked that []
built at 128621411. The verification run was:
[]
but the same target failed to build at 128621412. The sponge link for this run is available here: []
To see all targets that ran, along with their final status, visit: []Questions? Comments? See the URL.
[]

*** Original change description ***

Introduce RequestFulfillment as a way to model the operation that turns the logical concept of a binding into a CodeBlock suitable for preforming the operation requested by a DependencyRequest.

This is a very basic operation refactoring right now.  The follow-on operation will be to reorganize AbstractComponentWriter so that we can look at all DependencyRequests for a given BindingKey and then select the best, most efficient implementation of RequestFulfillment to use throughout.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128635514